Create SpellCheck.sh
Description : This script takes in a text file, a dictionary file (a text file containing a large list of words), and an integer provided by user to determine the number of most common typos to display. Punctuation, quotes, and numerals will be ignored. All words in the provided text file will be compared to the words in the dictionary file, and if the words are not in the dictionary, then they will be considered a typo. Single apostrophes are not used as delimiters since they could be a part of contractions. More complex grammar mistakes will not be considered.